item can be deleted from front end

+++ app/views/item.py
@@ -121,15 +121,23 @@ def editItem(category_id, item_id):
 @item.route('/category/<int:category_id>/items/<int:item_id>/delete',
 methods=['POST'])
 def deleteItem():
+    # grab a reference to the category and item models
     item = session.query(Item).filter(Item.id == item_id).one()
+    category = session.query(Category).filter(Category.id == category_id).one()
 
     try:
-        # delete item from the database
-        session.delete(item)
+        # remove category from item.
+        item.item_children.remove(category)
+
+        # if item doesn't have any categories then delete item from the database
+        if len(item.item_children) == 0:
+            session.delete(item)
+
+        # update the database
         session.commit()
 
         # send feedback to the user
-        flash("Successfuly deleted %s" % item.name)
+        flash("Successfuly deleted %s from %s" % item.name, category.name)
 
         return redirect(url_for("category.allCategories"))
